Move GameOptions and DEFAULT_GAME_OPTIONS to their own class.
Anything to make Game.ts smaller. :D
