+++ src/Game.ts
@@ -49,7 +49,6 @@ import {Turmoil} from './turmoil/Turmoil';
 import {RandomMAOptionType} from './common/ma/RandomMAOptionType';
 import {AresHandler} from './ares/AresHandler';
 import {IAresData} from './common/ares/IAresData';
-import {AgendaStyle} from './common/turmoil/Types';
 import {GameSetup} from './GameSetup';
 import {CardLoader} from './CardLoader';
 import {GlobalParameter} from './common/GlobalParameter';
@@ -68,99 +67,12 @@ import {AddResourcesToCard} from './deferredActions/AddResourcesToCard';
 import {isProduction} from './utils/server';
 import {ColonyDeserializer} from './colonies/ColonyDeserializer';
 import {GameLoader} from './database/GameLoader';
+import {DEFAULT_GAME_OPTIONS, GameOptions} from './GameOptions';
 
 export interface Score {
   corporation: String;
   playerScore: number;
 }
-
-export interface GameOptions {
-  boardName: BoardName;
-  clonedGamedId: GameId | undefined;
-
-  // Configuration
-  undoOption: boolean;
-  showTimers: boolean;
-  fastModeOption: boolean;
-  showOtherPlayersVP: boolean;
-
-  // Extensions
-  corporateEra: boolean;
-  venusNextExtension: boolean;
-  coloniesExtension: boolean;
-  preludeExtension: boolean;
-  turmoilExtension: boolean;
-  promoCardsOption: boolean;
-  communityCardsOption: boolean;
-  aresExtension: boolean;
-  aresHazards: boolean;
-  politicalAgendasExtension: AgendaStyle;
-  solarPhaseOption: boolean;
-  removeNegativeGlobalEventsOption: boolean;
-  includeVenusMA: boolean;
-  moonExpansion: boolean;
-  pathfindersExpansion: boolean;
-
-  // Variants
-  draftVariant: boolean;
-  initialDraftVariant: boolean;
-  startingCorporations: number;
-  shuffleMapOption: boolean;
-  randomMA: RandomMAOptionType;
-  soloTR: boolean; // Solo victory by getting TR 63 by game end
-  customCorporationsList: Array<CardName>;
-  cardsBlackList: Array<CardName>;
-  customColoniesList: Array<ColonyName>;
-  requiresMoonTrackCompletion: boolean; // Moon must be completed to end the game
-  requiresVenusTrackCompletion: boolean; // Venus must be completed to end the game
-  moonStandardProjectVariant: boolean;
-  altVenusBoard: boolean;
-  escapeVelocityMode: boolean;
-  escapeVelocityThreshold?: number;
-  escapeVelocityPeriod?: number;
-  escapeVelocityPenalty?: number;
-}
-
-export const DEFAULT_GAME_OPTIONS: GameOptions = {
-  altVenusBoard: false,
-  aresExtension: false,
-  aresHazards: true,
-  boardName: BoardName.ORIGINAL,
-  cardsBlackList: [],
-  clonedGamedId: undefined,
-  coloniesExtension: false,
-  communityCardsOption: false,
-  corporateEra: true,
-  customColoniesList: [],
-  customCorporationsList: [],
-  draftVariant: false,
-  escapeVelocityMode: false, // When true, escape velocity is enabled.
-  escapeVelocityThreshold: constants.DEFAULT_ESCAPE_VELOCITY_THRESHOLD, // Time in minutes a player has to complete a game.
-  escapeVelocityPeriod: constants.DEFAULT_ESCAPE_VELOCITY_PERIOD, // VP a player loses for every `escapeVelocityPenalty` minutes after `escapeVelocityThreshold`.
-  escapeVelocityPenalty: constants.DEFAULT_ESCAPE_VELOCITY_PENALTY,
-  fastModeOption: false,
-  includeVenusMA: true,
-  initialDraftVariant: false,
-  moonExpansion: false,
-  moonStandardProjectVariant: false,
-  pathfindersExpansion: false,
-  politicalAgendasExtension: AgendaStyle.STANDARD,
-  preludeExtension: false,
-  promoCardsOption: false,
-  randomMA: RandomMAOptionType.NONE,
-  requiresMoonTrackCompletion: false,
-  removeNegativeGlobalEventsOption: false,
-  requiresVenusTrackCompletion: false,
-  showOtherPlayersVP: false,
-  showTimers: true,
-  shuffleMapOption: false,
-  solarPhaseOption: false,
-  soloTR: false,
-  startingCorporations: 2,
-  turmoilExtension: false,
-  undoOption: false,
-  venusNextExtension: false,
-};
 
 export class Game {
   // Game-level data
