+++ src/SerializedGame.ts
@@ -8,7 +8,7 @@ import {SerializedPlayer} from './SerializedPlayer';
 import {SerializedDealer} from './SerializedDealer';
 import {SerializedTurmoil} from './turmoil/SerializedTurmoil';
 import {PlayerId, GameId, SpectatorId} from './common/Types';
-import {GameOptions} from './Game';
+import {GameOptions} from './GameOptions';
 import {IAresData} from './common/ares/IAresData';
 import {LogMessage} from './common/logs/LogMessage';
 import {SerializedBoard} from './boards/SerializedBoard';
