+++ src/boards/OriginalBoard.ts
@@ -6,7 +6,7 @@ import {ISpace} from './ISpace';
 import {BoardBuilder} from './BoardBuilder';
 import {SerializedBoard} from './SerializedBoard';
 import {Random} from '../Random';
-import {GameOptions} from '../Game';
+import {GameOptions} from '../GameOptions';
 
 export class OriginalBoard extends Board {
   public static newInstance(gameOptions: GameOptions, rng: Random): OriginalBoard {
