+++ src/database/IDatabase.ts
@@ -1,4 +1,5 @@
-import {Game, GameOptions, Score} from '../Game';
+import {Game, Score} from '../Game';
+import {GameOptions} from '../GameOptions';
 import {GameId, PlayerId, SpectatorId} from '../common/Types';
 import {SerializedGame} from '../SerializedGame';
