+++ src/database/PostgreSQL.ts
@@ -1,5 +1,6 @@
 import {IDatabase} from './IDatabase';
-import {Game, GameOptions, Score} from '../Game';
+import {Game, Score} from '../Game';
+import {GameOptions} from '../GameOptions';
 import {GameId, PlayerId, SpectatorId} from '../common/Types';
 import {SerializedGame} from '../SerializedGame';
 import {Pool, ClientConfig} from 'pg';
