+++ src/routes/Game.ts
@@ -6,7 +6,8 @@ import {BoardName} from '../common/boards/BoardName';
 import {RandomBoardOption} from '../common/boards/RandomBoardOption';
 import {Cloner} from '../database/Cloner';
 import {GameLoader} from '../database/GameLoader';
-import {Game, GameOptions} from '../Game';
+import {Game} from '../Game';
+import {GameOptions} from '../GameOptions';
 import {Player} from '../Player';
 import {Server} from '../models/ServerModel';
 import {ServeAsset} from './ServeAsset';
