+++ tests/cards/community/ExecutiveOrder.spec.ts
@@ -2,7 +2,8 @@ import {expect} from 'chai';
 import {Player} from '../../../src/Player';
 import {cast, setCustomGameOptions} from '../../TestingUtils';
 import {TestPlayers} from '../../TestPlayers';
-import {Game, GameOptions} from '../../../src/Game';
+import {Game} from '../../../src/Game';
+import {GameOptions} from '../../../src/GameOptions';
 import {PartyName} from '../../../src/common/turmoil/PartyName';
 import {ExecutiveOrder} from '../../../src/cards/community/ExecutiveOrder';
 import {SelectPartyToSendDelegate} from '../../../src/inputs/SelectPartyToSendDelegate';
